requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.0

# OpenAI Integration
openai==1.10.0
tiktoken==0.5.2

# NLP & Embeddings
sentence-transformers==2.3.1
hazm==0.7.0  # Persian NLP library
parsivar==1.0.0  # Persian text processing

# Vector Store & RAG
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4

# Graph & Visualization
networkx==3.2.1
plotly==5.18.0
kaleido==0.2.1  # Static image export

# Data Processing
pandas==2.1.4
pydantic==2.5.3  # Data validation
pydantic-settings==2.1.0  # For settings management
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
jdatetime==4.1.1  # Persian calendar support
ratelimit==2.2.1  # Rate limiting for API calls
